Adding Deleting Getting nearly finalized. Control segments added
Some Libraries and Managers added to handle some cases.

Claim and Policy is removed due to  it is unnecessary to this project.

Errors when adding book with turkish characters to database is fixed.

Nearly All Delete functions have been added to project.

To Book Controller:
-Removing Book from all books in database.
-Removing UserBooks from UserBooks in database.
Added.

To UserController:
-Adding to UserBook has some controls when adding (same book/no book cant added).
-Getting user's favourite books has some controls now (no book).
-Adding to Favourite have some controls now (no book/same book can't be added).
-Delete From User books and Remove from favourites have been added with controls(no book).

User can't add same book to his favourites or his books (UserBook).